requirements.txt
pywebview>=4.4
fastapi>=0.111
uvicorn[standard]>=0.29
faster-whisper>=1.0
static-ffmpeg>=2.5
httpx>=0.27
psutil>=5.9
platformdirs>=4.2
rich>=13.0
python-multipart>=0.0.9
